FormFinty.php
The above code is a registration form for yoga classes. It starts with a title and a style tag to give a background color to the page and font family. The body of the code has a <table> tag with a class form-container. This table has two headings and a form. The form has an action attribute which links the form to a new page, called register.php. The form also has a method attribute, which states that the data should be sent using the POST method. 

The table has a form-control class which helps to style the text box and drop-down list. The form-label class is used to style the labels of the text box and drop-down list. The form also has a submit button which is styled using the form-btn class.

The last block of code is to insert the data into the database. The code starts with setting the variables to the values obtained from the form. Then the connection to the database is created. After the connection is established, an SQL query is used to insert the data into the database. The values are taken from the variables. If the query is successful, it returns a message saying the record was created successfully, else it returns an error message. The connection is then closed.

The form also has a JavaScript function which is used to validate the form. The function has two conditionals which check if the age is between 18 and 65 and if the payment is at least 500 INR. If the conditions are not met, an alert is returned. If the conditions are met, the form is submitted.